# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/uploads": {
      "post": {
        "operationId": "MediaController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/png": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/webp": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "The request body is the file itself (e.g. `Content-Type: image/png`), not multipart: behind CloudFront OAC the\nclient must send the body's SHA-256, easy for raw bytes. The stored type still comes from the magic bytes.",
        "tags": [
          "admin/media"
        ]
      }
    },
    "/admin/media": {
      "get": {
        "operationId": "MediaController_list",
        "parameters": [
          {
            "name": "unused",
            "required": false,
            "in": "query",
            "description": "Only media not referenced by the profile or any project.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "mime",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "image/jpeg",
                "image/png",
                "image/webp"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
    "/admin/media/{id}": {
      "delete": {
        "operationId": "MediaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/media"
        ]
      }
    },
...
    "/admin/skill-categories": {
      "get": {
        "operationId": "SkillCategoriesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skill-categories"
        ]
      },
      "post": {
        "operationId": "SkillCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skill-categories"
        ]
      }
    },
    "/admin/skill-categories/reorder": {
      "patch": {
        "operationId": "SkillCategoriesController_reorder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skill-categories"
        ]
      }
    },
    "/admin/skill-categories/{id}": {
      "get": {
        "operationId": "SkillCategoriesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skill-categories"
        ]
      },
      "patch": {
        "operationId": "SkillCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSkillCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skill-categories"
        ]
      },
      "delete": {
        "operationId": "SkillCategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skill-categories"
        ]
      }
    },
    "/admin/skills": {
      "get": {
        "operationId": "SkillsController_list",
        "parameters": [
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skills"
        ]
      },
      "post": {
        "operationId": "SkillsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skills"
        ]
      }
    },
    "/admin/skills/reorder": {
      "patch": {
        "operationId": "SkillsController_reorder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skills"
        ]
      }
    },
    "/admin/skills/{id}": {
      "get": {
        "operationId": "SkillsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skills"
        ]
      },
      "patch": {
        "operationId": "SkillsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSkillDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skills"
        ]
      },
      "delete": {
        "operationId": "SkillsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/skills"
        ]
      }
    },
...
    "/admin/services": {
      "get": {
        "operationId": "ServicesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/services"
        ]
      },
      "post": {
        "operationId": "ServicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/services"
        ]
      }
    },
    "/admin/services/reorder": {
      "patch": {
        "operationId": "ServicesController_reorder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/services"
        ]
      }
    },
    "/admin/services/{id}": {
      "get": {
        "operationId": "ServicesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/services"
        ]
      },
      "patch": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/services"
        ]
      },
      "delete": {
        "operationId": "ServicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/services"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "admin": {
        "type": "apiKey",
        "in": "header",
        "name": "x-authorization",
        "description": "Bearer <access token>"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "LocalizedTextDto": {
        "type": "object",
        "properties": {}
      },
...
      "ReorderDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateSkillCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "visible": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateSkillCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "visible": {
            "type": "boolean"
          }
        }
      },
      "CreateSkillDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Technology name; not translated.",
            "maxLength": 60
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "visible": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "categoryId"
        ]
      },
      "UpdateSkillDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Technology name; not translated.",
            "maxLength": 60
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "visible": {
            "type": "boolean"
          }
        }
      },
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "title": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "description": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "visible": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "description"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "title": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "description": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "visible": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}